backend/requirements.txt
alembic==1.13.2
apispec==6.6.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-redis==0.4.0
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
SQLAlchemy==2.0.32
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webargs==8.4.0
Werkzeug==3.0.3
WTForms==3.1.2
